"BigDecimal changed to Long to preserve coincide with CLP format. Code refactored accordingly"

+++ ms-reports/src/main/java/com/quandrix/ms_reports/dto/SalesReportResponse.java
@@ -1,6 +1,5 @@
 package com.quandrix.ms_reports.dto;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import lombok.AllArgsConstructor;
@@ -12,6 +11,6 @@ public class SalesReportResponse {
     private LocalDateTime from;
     private LocalDateTime to;
     private long totalTransactions;
-    private BigDecimal totalAmount;
-    private BigDecimal averageAmount;
+    private Long totalAmount;
+    private Long averageAmount;
 }
